Finished prototype.
-added pause state
-added game over state
-added score
-added button sounds
-balanced levels
-fixed all warnings
-removed getComponent bottleneck (O(n) -> O(1))

+++ Tower Defense/AISystem.cpp
@@ -19,7 +19,7 @@ AISystem::AISystem()
 		sf::RectangleShape shape;
 		shape.setSize({ 10, 10 });
 		shape.setOrigin(5, 5);
-		shape.setPosition(x, y);
+		shape.setPosition(static_cast<float>(x), static_cast<float>(y));
 
 		checkpoints.push_back(shape);
 	}
